fix(flutter): Always strip internal sync span marker
The ThreadInfoIntegration handlers only stripped the internal 'sync'
marker when it equaled true, so a stray non-true value would leak to
Sentry. Gate on marker presence and clear it unconditionally, while
still setting blocked_main_thread only when the span is actually
synchronous.

Also move the synchronous-marker extensions and their shared private
key out of instrumentation_span.dart into a dedicated
synchronous_span_marker.dart, so the raw key stays in one place. No
public API or v1 behavior change.
