+++ packages/dart/lib/src/tracing/instrumentation/instrumentation.dart
@@ -3,3 +3,4 @@ library;
 
 export 'instrumentation_span.dart';
 export 'span_factory.dart';
+export 'synchronous_span_marker.dart';
